services/crawler-service/pyproject.toml
[tool.poetry]
name = "crawler_service"
version = "0.1.0"
description = ""
authors = ["Nguyen Tran <merlinvn@gmail.com>"]
readme = "README.md"
packages = [{ include = "crawler_service", from = "." }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
openai = "^1.51.2"
pydantic = "^2.9.2"
aio-pika = "^9.4.3"
pydantic-settings = "^2.6.0"
httpx = "^0.27.2"
crawl4ai = "^0.3.71"
setuptools = "^75.2.0"
google-generativeai = "^0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "crawler_service.main:start"
